fix: compilation errors in ObexClient + AVRCP
- Merge duplicate companion objects in ObexClient (two -> one)
- Add MAX_CONTINUATIONS to existing companion object
- Fix Int->Byte type mismatches: OP_CONNECT, OP_DISCONNECT, OP_SET_PATH, OBEX_VERSION, OBEX_FLAGS
- Fix Int->Long mismatch in AVRCP UID parsing (or with Long needs Long operand)
Resolves all compilation errors that blocked PR #431 CI.

+++ app/src/main/java/com/btsec/testtool/data/bredr/AvrcpSecurityRepositoryImpl.kt
@@ -305,7 +305,7 @@ class AvrcpSecurityRepositoryImpl
                 // Extract UID (8 bytes, big-endian)
                 var uid = 0L
                 for (i in 0 until 8) {
-                    uid = (uid shl 8) or (data[offset + i].toInt() and 0xFF)
+                    uid = (uid shl 8) or ((data[offset + i].toInt() and 0xFF).toLong())
                 }
                 offset += 8
